requirements.txt
flask
google-auth-oauthlib==0.4.6
google-auth==2.3.1
requests-oauthlib==1.3.0
google==3.0.0
cachecontrol==0.12.11
requests==2.26.0
flask-bcrypt
flask_sqlalchemy
flask-cors
flask-sqlalchemy==3.0.2
flask-jwt-extended
authlib
pyOpenSSL==19.0.0
flask-login
flask-migrate
psycopg2-binary
postgres
bulma
python-dotenv
flask_wtf
flask-restx
# verificar que FLASK_APP == project 
# export FLASK_APP=project
# export FLASK_DEBUG=1

#psql -h localhost -U admin -d cellwatch